Extract redacting to a `Redactor`, override `PierLogging::Logger`s log method
